Preparation for MAP v4.1 release (#74)
code improvement to deal with join mismatch error when no ICEs are detected in the sample
PROKKA is not part of the pipeline
AMRfinderplus is no part of the pipeline. It will be back soon though.
Mobile-OG annotation is not supported. I will be back soon as well.
An end-to-end nf-test is in place
The pipeline supports compressed input and outputs